This commit moves the event generation code to separate file in controller package
