+++ pkg/controller/pv_controller.go
@@ -23,7 +23,6 @@ import (
 	"time"
 
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	corev1informer "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes"
@@ -55,14 +54,6 @@ type PVEventController struct {
 
 	// Recorder is an event recorder for recording Event resources to Kubernetes API.
 	recorder *Recorder
-}
-
-// Recorder is a wrapper over EventRecorder which helps to
-// control event generation
-type Recorder struct {
-	record.EventRecorder
-	// generateEvents will control the event generation based on its value
-	generateEvents bool
 }
 
 // NewPVEventController will create new instantance of PVEventController
@@ -141,28 +132,4 @@ func GetSyncInterval() time.Duration {
 		return sharedInformerInterval
 	}
 	return time.Duration(resyncInterval) * time.Second
-}
-
-// Event is a wrapper over original Event which will help to generate events
-func (r *Recorder) Event(object runtime.Object, eventtype, reason, message string) {
-	if !r.generateEvents {
-		return
-	}
-	r.EventRecorder.Event(object, eventtype, reason, message)
-}
-
-// Eventf is wrapper over original Eventf, it is just like Event, but with Sprintf for the message field.
-func (r *Recorder) Eventf(object runtime.Object, eventtype, reason, messageFmt string, args ...interface{}) {
-	if !r.generateEvents {
-		return
-	}
-	r.EventRecorder.Eventf(object, eventtype, reason, messageFmt, args...)
-}
-
-// AnnotatedEventf is wrapper over original AnnotatedEventd just like eventf, but with annotations attached
-func (r *Recorder) AnnotatedEventf(object runtime.Object, annotations map[string]string, eventtype, reason, messageFmt string, args ...interface{}) {
-	if !r.generateEvents {
-		return
-	}
-	r.EventRecorder.AnnotatedEventf(object, annotations, eventtype, reason, messageFmt, args...)
 }
